refactor: change AuthService to authenticate util
