+++ src/app/api/workflow-states/workflowStates.controller.ts
@@ -1,11 +1,11 @@
 import { NextRequest, NextResponse } from 'next/server'
-import AuthService from '@api/core/services/auth.service'
 import WorkflowStatesService from '@api/workflow-states/workflowStates.service'
 import { CreateWorkflowStateRequestSchema } from '@/types/dto/workflowStates.dto'
 import httpWorkflowState from 'http-status'
+import authenticate from '@api/core/utils/authenticate'
 
 export const getWorkflowStates = async (req: NextRequest) => {
-  const user = await AuthService.authenticate(req)
+  const user = await authenticate(req)
 
   const workflowStatesService = new WorkflowStatesService(user)
   const workflowStates = await workflowStatesService.getAllWorkflowStates()
@@ -14,7 +14,7 @@ export const getWorkflowStates = async (req: NextRequest) => {
 }
 
 export const createWorkflowStates = async (req: NextRequest) => {
-  const user = await AuthService.authenticate(req)
+  const user = await authenticate(req)
 
   const data = CreateWorkflowStateRequestSchema.parse(await req.json())
   const workflowStatesService = new WorkflowStatesService(user)
